camps/camp2-gateway/samples/mcp-custom-headers/pyproject.toml
[project]
name = "mcp-custom-headers-sample"
version = "0.1.0"
description = "Sample demonstrating MCP custom headers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.15.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
agent-framework = [
    "agent-framework>=1.0.0b250000",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
